pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "VATSAL AI Desktop Automation - Python 3.11+ Compatible"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=43.0.3",
    "psutil>=6.1.0",
    "pyautogui>=0.9.54",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.0.1",
    "pyttsx3>=2.98",
    "pywhatkit>=5.4",
    "requests>=2.32.3",
    "sift-stack-py>=0.9.1",
    "speechrecognition>=3.14.0",
    "watchdog>=6.0.0",
    "Flask>=3.1.2",
    "flask-cors>=6.0.1",
    "Flask-SocketIO>=5.5.1",
    "python-socketio>=5.14.3",
    "selenium>=4.38.0",
    "beautifulsoup4>=4.14.2",
    "pandas>=2.3.3",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "streamlit>=1.40.2",
    "pygame>=2.6.1",
]
